backend/requirements-edge.txt
# Edge Device Optimized Requirements
# Use this for Jetson Nano or similar low-resource devices

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (lightweight)
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis (minimal)
redis==5.0.1

# AI - Offline Only (no OpenAI dependency)
# openai==1.3.5  # Comment out for pure offline
langchain==0.0.340
langchain-community==0.0.10
tiktoken==0.5.2

# Offline Voice (CPU optimized)
openai-whisper==20231117
pyttsx3==2.90
# Use CPU-only PyTorch for Jetson Nano
# torch==2.1.0+cpu  # Install separately based on your system
# torchaudio==2.1.0+cpu

# HTTP Client for Ollama
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Processing (lightweight)
PyPDF2==3.0.1
Pillow==10.1.0

# AWS (optional, for cloud storage)
# boto3==1.29.7  # Comment out if not using AWS

